Part Nine: update for other views
Update Edit, Delete and Details views, including to update the binding
statement in the controller for Edit.  Also decided to change the date
format in the Movie class, as the {0: d} version was somehow causing the
date field edit form to not have the existing date data persist ,
causing the user to re-enter the release date for every edit action.